[AD-436] Generalize Glue and Face and port Vty
